Revise and add links to the shorthands.
Some of them needed tweaking, either for precision or to match the changes to the spec between domenic/promises-unwrapping and @allenwb's draft. For example, ResolvePromise was inlined into Promise Resolve Functions, so I used that as an opportunity to make the "saving" of the `resolve` and `reject` functions explicit.

Added links to the ES spec draft, closing #4.